adds form, editProfileHandler to enable/disable profile edits
