+++ src/components/ExploreContainer.css
@@ -22,3 +22,7 @@
 .container a {
   text-decoration: none;
 }
+
+#profile-edit-icon {
+  
+}
