--- src/pages/Profile/Profile.css
@@ -1,6 +1,0 @@
-.avatar {
-  margin: auto;
-  height: 100px;
-  width: 100px;
-  margin-bottom: 30px;
-}
